fix(ppt): _fit_text 截斷分支 TypeError——組版整個失敗
實機 job 274 組版 exit=1：

    return text[: max(1, budget - 1)]...
    TypeError: slice indices must be integers

`budget = per_line * lines`，而 `per_line` **刻意是浮點數**——_text_capacity 的
註解寫著「向下取整等於白丟將近一個字的寬度」。浮點是對的，拿去當 slice index
必須先取整。

⚠ **為什麼到今天才爆**：`len(text) <= budget` 那條路徑不碰切片，而在此之前所有
文字都裝得下。直到 KP 頁尾註記變長（加了定位分類推導規則），第一次走進截斷分支
才現形，而後果是整份 PPT 產不出來。

這是「happy path 全綠、例外路徑從未執行」的典型：_fit_text 被呼叫過無數次，
但截斷那一半是死的。

regression test 已驗證能抓到原 bug（暫時還原成 float 版本 → 3 支立刻紅）。

+++ skills/patent-report-ppt/scripts/build_ppt.py
@@ -623,7 +623,11 @@ def _text_capacity(theme: Theme, *, width_in: float, height_in: float, size_pt: 
 def _fit_text(theme: Theme, text: str, *, width_in: float, height_in: float, size_pt: float) -> tuple[str, bool]:
     """把文字截到框內裝得下，超出加「…」。回傳（文字, 是否截斷）。"""
     per_line, lines = _text_capacity(theme, width_in=width_in, height_in=height_in, size_pt=size_pt)
-    budget = per_line * lines
+    # ⚠ `per_line` 刻意是浮點（見 _text_capacity：向下取整會白丟近一個字寬），
+    # 但拿來當 slice index 必須先取整——否則走到截斷分支就 TypeError。
+    # 2026-08-10 實機爆掉：KP 頁尾註記變長，第一次走到這個分支才現形，
+    # 在那之前所有文字都裝得下，`len(text) <= budget` 直接回傳。
+    budget = int(per_line * lines)
     if len(text) <= budget:
         return text, False
     return text[: max(1, budget - 1)].rstrip("，、。；：") + "…", True
